Add files via upload
Tasks were based on Automation Testing. I had to automate the testing procedure as it was shown in the provided Test Document. This was my first ever use of Cypress as a testing tool. One will find flaws in the existing code files but being new to Cypress, I tied to give my best and have already started exploring in more detail.